Round indicators for auction and functionality implemented in auction game manager. Draw indicator for auction.
4 round indicators created for auction in the auction game manager since the turns / rounds work differently for this game.
New functions for widgets in the auction game manager including: update turn indicators, setup turn indicators and update draw indicator.
Also new events in the auction game manager: Fade Out Turn indicator and Fade Out Draw Indicator. These use some pre-existing code from other fade out timelines but slightly different for the turn indicators since we do all of them with the corresponding player active bools.

Setup turn indicators moves the turn indicators to their correct positions and this is called in the setup auction function.
Update turn indicators sets the turn indicators with corresponding player active true to be visible and if we have any inactive players for the round then we call the fade out turn indicators event.
This function is called in the end turn function if we have a draw and is called in the next round function as well.

Update draw indicator function sets the location and rotation of the draw widget component so it is above the auction main card pile and is facing the player, this function also sets the widget visible and then calls the fade out draw indicator.
This function is called in the end round function when a draw is found to be true.