Auction unique card setup and some functionality.
New function in auction game manager called unique card action which has two passed parameters: card played which is of type unique card names and then also an integer for player index.
Called this function in the end round function after checking if the current card adding to the score is higher than 13 and then we set the result of the function to be the score of the current looping player.

Some new variables for unique cards but those are not fully implemented yet.
Joker, devil, famine, death implemented.