Auction further development of unique card interactions.
Unique card interactions fully implemented need to properly test but pretty sure they are all in and working for both the players and enemies.

Conquest and war is in current hand component as well as the auction game manager. Conquest allows the player to play a second card in the round so this changes how many cards are checked for in the select card function in the current hand component.

War is also setup in the select card function in the current hand component for players where the maximum number of cards is -1 so that they cannot select any however since NPCs directly set their selected cards we instead toggle whether they are active or not in the auction game manager.

New clear unique card checks function which handles when unique card checks should be active or not and is called at the start of the end round function. However the conquest and angel checks are also cleared at the end of the drawing functionality in the end round function.

New variable for current round index in the auction game manager this is not updated when drawing and is only updated after ending round.

Death now only gives one card to each player from the played card piles instead of dealing every card from those piles.

Debugging print strings in the unique card interaction function for all the unique cards.