Rules manager and some placeholder widgets. Changed move characters functions in the auction and IDI game managers.
New manager actor for the rules.
New functionality in this actor and in the game mode for showing and hiding the rules.
New widgets for the rules.
New widget for the rule text holder which has a multi line text variable that can be adjusted to present different rules in the holder.

Changed some of the calls in the setup and start game functions for both auction and i doubt it so that they are more easily tweakable for using the rules when you first "start" a game.

New functionality in auction for pausing the game in case it is a NPCs turn and the player can't play we now store the end round function call timer and can pause and unpause it in the corresponding functions which is done when showing and hiding the  rules for auction.

One new input action for hiding and showing the rules called toggle rules with functionality setup in the player to call the corresponding function in the rules manager for showing or hiding the rules.