Jukebox functionality and interaction implementation.
Using the jukebox actor.
New widget interaction component on player.
New input action for player in bar input mapping context.
Using widget Jukebox.
Added in all the music created for the game upto now for placeholder.
Created sound cues for all of these.
Created sound attenuation for these.
Applied sound attenuation for all these.
Jukebox actor:
static mesh - using new collision object type jukebox
collision - using new jukebox interaction collision preset
audio component
jukebox UI

Collision events for player entering and exiting the collision which updates a variable in the player as well that they are in the range of the jukebox.
Also activates and deactivates the widget interaction component so it only works when in the range of the jukebox.
Also starts the line of sight events in the jukebox.

Line of sight using a box trace for jukebox objects from the player's eyes using their control rotation forward.
If in line of sight then show UI.
If not then hide UI.

On begin play get reference to player and setup event for updating the audio time using event with audio component to get the playback percentage every tick.

Next song and previous song function just updating the song index based on the last index of the songs array.
New variables in the jukebox actor: audio time float,. songs sound cue array, current song index integer, player ref actor reference and player in range bool.

Pause/unpause song event using a multi gate to currently stop the audio component and play the audio component using the audio time for the start time so that it resumes at the correct point in the song.

Input action in player on left click checks if can interact with jukebox and if so then call press pointer key in the jukebox widget interaction so can click the jukebox diegetic UI buttons and then on canceled and completed release pointer key with widget interaction component.

Jukebox UI just has a jukebox reference setup in construct event and then on pressed events with the buttons to call the corresponding events in the jukebox for the button.