Began the AI for Auction
Working specifically on expert difficulty, but before difficulty is considered, all the cards that are in the NPC's hand are gotten and added to their corresponding arrays which are High Value, Mid Value, Low Value, and Unique.

Then we get the last card played for Joker interaction

We then check what the card up for auction is with a switch statement, and I am currently working on the algorithm for when an Ace is up. Some progress has been made for this, but will expand on next commit when complete.