Jukebox widget animations. Jukebox fixes.
Jukebox widget animations for hovering and for clicking implemented.

Jukebox implementation for fading in and fading out audio for pausing, resuming, next and previous song and also for the initial starting of the audio on begin play.
2 new variables for time for fade in and time for fade out.

Fixed an issue with the audio time tracking by turning off auto activate on the audio component and then also doing fade in on begin play.