Outside looping functionality. Organised jukebox variables.
2 new actors for outside looping mechanic.
Outside loop collision which includes the functionality for this mechanic.
Outside loop position which is an actor with a directional arrow in it which is used for setting the player's position and control rotation to this actor's location and the rotation.

Functionality for looping collision: overlap with collision which only overlaps with player which then starts a camera fade with time as the variable fade time and fades to black with hold when finished active. Then we delay for duration variable time before moving and fading in where we then set the player control rotation and location to the rotation and location of the set loop end position actor and then we start a camera fade to alpha 0 with time fade time variable value.

Organised jukebox variables into categories.