Jukebox tweaked press animations. Jukebox fix. Main Menu Functionality.
Jukebox press animations changed so more satisfying.

Jukebox fixed with changing song when paused and then trying to pause would then resume instead.
Now using a bool for song paused instead of using a multigate since its just easier to track when the audio is paused or not.

Main menu actor.
Contains 2 splines: one for credits and one for exiting the game.
Interact collision for making the UI appear and also starting the move to main function.
Main SM - static mesh of the thing that the main menu will be on.
Credits start location and rotation sphere with no collision and an arrow for direction attached.
Main location and rotation sphere with no collision and an arrow for direction attached as well as the actual main menu widget.
Options location and rotation sphere ... which also has the options widget attached to it as well.

Variables:
time to main
time to options
time to credits
time to exit

player ref

start position before moving
start rotation before moving

credits timeline active.

Multiple event graphs to split everything up.

Begin play get reference to player and move the credits start location and rotation to be the location and rotation of the credits spline at distance 0.

Most of the event graphs do the same thing with slight differences.
Main functionality includes setting the start position and rotation before moving variables.
Then setting ignore look and move input.
Set play rate based on the event graph e.g. using time to main for the to main graph.
Then we use a timeline to lerp the position and rotation of the player to the position of the corresponding thing depending on the graph.
Setting control rotation instead of just actor rotation since it is the player and their view needs to change.
Then on finished we do various actions depending on the event graph.
For main we set the main menu widget visible and then set the can interact with main menu bool in the player true and then activate the widget interaction component in the player and set the ignore look input to be false so the player can look around to interact with the menus.

For the options graph on timeline completed we set the basically do the same thing but with setting the options widget visible since it is not visible by default and is also set to be not visible when moving to main.

For credits we call the move player along credits spline with the time to finish currently being 10 but this can be changed at any time in development.

For quit we just call the move player along exit spline with a time to finish of 3 seconds which can be changed at any time during development and then we also starta  camera fade to black with a current duration of 2.8 seconds and with hold when finished active.

In the exit spline movement and credits spline movement we first update the play rate of the timeline we are about to use based on the passed time to finish variable and then we lerp the rotation and location of the player's location and control rotation based on the distance along the spline by using the length of the spline * the alpha of the timeline and then on finished for the quit spline we call quit game with no specified player.
With the finished credits spline timeline we check if the credits timeline active bool is true and if so we call this event again with the same time to finish.
There is another event in the credits spline movement graph to stop this timeline and return to the main menu proper.