Auction AI progression
Created local variable Won Cards Map of Enum Card Names with integer.
This Map is used at when "auction decide" is called, and adds to the integer based on how many cards have been won.

Decision making for Jack is complete, first checks if there are any unique cards in the hand, if there is random bool by weight 0.5 to randomise whether a unique is played or not. If false, play the lowest possible card.

Jack Unique Options:

Devil: Checks if an ace has been won, if it has play the card through the break. If not, check if any cards have bene won. If they have, see if the NPC makes a mistake and plays the devil anyway with bool with weight

War, Death - Select if no other option

Pestilence - If no aces are won, play this card on break, if there have been, select card but do not break, if all aces played, continute loop

Conquest, Famine - Break the loop and auto play these cards

Angel - Do not play

Ace decision making:

Try to get a high value card, then check if we have a joker. If we do, is the last card played a higher value than the card we selected? If it is, pick joker. If not pick the other card

If there are no high value cards, get a mid value card, and if that isn't possible, try and get a unique.

Ace unique options

Joker - Checks if Joker is a bigger value than all other cards, if it is then play it and break loop

War, Death, Pestilence - Low priority unique, if another unique has been selected, ignore.

Devil Functions the same as Jack

Famine, Angel - Both will automatically be selected and break the loop