Darts furthering. Dart throwing. Dart score point actor.
Darts furthered functionality a little more.
Dart throwing towards intended point actually working now.
Dart position on dart board actually working now.
Created an actor with some variables like score, treble?, double? to hold the score and to be used for getting the closest of these actors to determine the score of the dart.
Dart board holds an array of these actors need to actually add them all as child actors and set it all up and then finish the functionality and darts will be functionally done for now.