Darts working score calculation. Darts bust check.
Using get nearest actor and using the array of dart score point actors in the dart board actor to then get the closest of these actors. Then we get the score from this actor for the score of the dart hitting the board.
This all works as intended since I've added over 120 score point actors as child actors to the dart board which isn't the best but it works and isn't really performative heavy since they are just very small spheres that aren't rendered and only contain 3 variables: score integer, treble bool and double bool.
The values for these variables are set for the defaults in the dart board actor.
Using parent spheres that are also not rendered so that each row of score point actors make perfect circles.

New function in the dart game manager for bust check which is called in the dart play function after the game over check is false and basically checks if the passed score would make the player's score go below 0 and if it does then they bust and cannot play any more darts for the round and the next player's turn starts.