Darts scoreboard widget and functionality. Some dart board game manager fixes.
New widget for dart scoreboard which is added to the dart board.
Horizontal box with 2 vertical boxes within it and text for player names and scores and then 3 texts for each player's score for each of their darts for the round.
3 Functions: setup, change dart score, and update player score.
Variable: player names text array for now.
Setup takes player names text array and a bool called game type which is currently used to determine if the game played is going to be a 501 or 301 but not setup yet the difference between these games in the game manager nor how to select the different games.
We first set the variable player names to this passed array and then change the 2 texts for player names - scores to be the names and then either 501 or 301 depending.
We then set all of the dart scores to be empty texts.

Change dart score function is as follows:
pass dart number, score, player index and then a bool for if this dart has busted the player.
We then change the according dart text to these passed parameters.
We also set the second and third corresponding dart scores to be empty if we are changing the score of the first dart score since it is now a new round so these values need to be cleared.

Update player score passes player index, and new score and then changes the corresponding player score text to be the correct name and then new score.

These functions are called in the darts game manager using the dart board reference which now holds a direct reference to the scoreboard widget not the component.
2 new events in the dart board actor for hiding and showing the scoreboard which is called in the begin game function in the dart game manager and the leave / cancel game function in there as well.

The change dart score function is called in the play dart function passing all correct info using the score generated in this function.
New variable in the dart game manager called player started first bool which just shows if the player is the one who went first so we can determine the player index to pass to the scoreboard widget.
The update player score function is called in the end player turn function after calculating the new score of the player that is ending their turn.

Also fixed some issues with the dart board game manager calculating the player scores since I wasn't clearing the round scores after using them in the end player turn so the scores the player's were getting were cumulative instead of resetting after each round.