Added a check to queen and king auction and
Kings and queens now check for higher cards first before playing mid level cards. They check for eights and nines but dont play tens

Reset the "selected card" and new conquest active variable at the start of decision making. "ConquestActive?" is checked with the Auction game manager.

Pushing this because im gonna change the selected card variable to an array so doing this as a precaution in case i break something