Pestilence change. Draw round bool stored in auction game manager. Need to do bug fixing for auction seems like theres quite a few issues / bugs but some of these are the AI not being finished yet tbf.
Pestilence now only affects the players that haven't played the card.

New bool in the auction game manager that tracks whether the current round is a drawed round which is set after checking if the current round is a drawn round. This is used to check if the pestilence should be active for the round since pestilence doesn't activate for drawn rounds.

There's quite a few bugs with auction that needs to be fixed but think some of it is due to AI not being fully done yet. There needs to be some other testing as well