NPC interaction update to diegetic UI instead of button press. Overall diegetic UI widget interaction deactivation checks so works with overlapping actors.
NPC interaction is now based on a button that shows when the player is within the interaction range and the player must click on that the same way they do with the darts game and the jukebox. Not deleted the previous code for interaction but have disconnected it can always get rid of it before building however as well as the input action for interaction with NPCs.
This start game button rotates to face the player whilst it is visible.

Now when attempting to deactivate the widget interaction component we first check if we are overlapping with any other actors that also have diegetic UI so that we don't deactivate when we shouldn't and so that we can have multiple overlapping diegetic UI actors with interaction remaining correct and input remaining able when leaving only one of the overlapping actors.