Button prompt functionality setup in game mode. New entry in all game names enum - "none". Some changes the set current game mode and leave current game functions in the game mode. Camera fades for starting games. Changed credits so can look around whilst on the spline.
New widgets for all the games and for the bar which will be non-diegetic and will contain all of the button prompts for the games and the bar.
Functionality setup for adding and removing these from the screen in the game mode.
Setup and storing of the created widgets in the game mode as well.

New entry in the all game names enum for none since there wasn't one before and now can tweak some more stuff in the game mode for set current game mode and leave / cancel current game functions which are in this commit as well.

Camera fade for starting all games now implemented now using new function in the NPC which is called from the game availability button click event within the widget. The function just starts a camera fade lasting one second that is held on completion and then in the initiate game function in the NPC we do another camera fade from 1 to 0 alpha over 1 second not holding on completion. This initiate game function is called after the camera fade function with a 1 second timer in the same widget.

Changed the credits functionality so can now look around during the credits movement spline functionality since its probably better. This change is in the main menu actor.