Pause screen widget. Pause screen basic widget animations. Pause functionality in the game mode. Remove button prompt on pausing and adding it back when unpausing.
Started working on the pause screen widget functionality.
Hover, unhover and click functionality for each button in the pause menu.
Pause functionality for input action using player with input action in the player.
Pause functionality is actually in the game mode.

Remove the current button prompt when pausing and add it back when unpausing in the game mode.