Starting working on the zoom mechanic for the different games. New input action. Functionality in the player.
Will work on this more later theres going to be some bugs with it with darts specifically but some of the other games should be fine.
New input action for zooming in which is a down input action that starts a timeline and reverses a timeline when pressed and released.
More target points in the world with tags Zoom_In_GameName need to actually make all of these some of the names are shortened however for the tag.
Majority of the functionality for this is in the player. Would move this to a component but timelines are just easier realistically.
Different curves for X and Y versus Z so can have the player go up quicker than along.