Options diegetic menu functionality and look (temp). UI Click input action is now also down. Tried messing around with font to get it scaling properly but couldnt figure it out for now.
Basically put the pause screen options menu into the diegetic options menu and had to change around some of the things like not using a canvas panel and instead using a sizebox so that it would look correct in world.

UI click input action is now also a hold down so can hold down on the sliders in world space to change them. Only had to add the down method to the input action as well.

Tried to make the font that we're using be a vector based font or something like that to make it so that it would scale properly for world space UI but couldn't figure it out will try again another day.