Swapped around the button prompt ordering. Updated button prompt for zoom in. Zoom in functionality furthered. Target points for zoom in have different arrows (size and colour).
Button prompt ordering is now name of action followed by hyphen and then button prompt image.

Updated button prompts for zoom in.

Target points now have arrow components on them which are blue and slightly smaller so easier to orient.

Zoom in functionality furthered: zoom in with hiding away the NPC card hands for auction and I Doubt It. New functions in the card interface for these hiding animations. New checks in the update visible hand in the current hand component so not glitchy whilst zoomed in.

Darts hiding active sliders whilst zoomed in. Darts zoomed in changing position when moving characters and not moving the characters when zoomed in so no giltchyness.

Tried to setup various ways to pause the games without having to do a bunch more code and timers but none of them worked so right now pausing the game only really works when the turn is still in a timer. So we either take the time to change a lot of the code for pausing to be fully implemented or we just leave it but it also means that pausing the game when in a game won't fully pause the game for all the games as well.