Card holo representation functionality. Doubt showing cards functionality and aesthetics.
New actor for holo card.
New child actor on card piles.
New functions for hiding and showing the holo card.
When showing the holo card we pass some card info to it and then update the card model and such with this new card info.

Updated add card to pile functions to pass whether or not the holo rep should be shown.
Gone through each instance of add card to pile function calls and set whether they should show the holo version of the card.

New child actor of the card pile actor specifically for the cards that have just been played that is added to this card pile when doubting.
Changed the doubt correct and incorrect function calls to timers and made the time for these variables as well as the amount of time to resume the player turn after doubting has ended as a variable.
So there is now delays between doubting being called, doubting being checked and the game being resumed. All to give time for the player to see and determine whether or not the doubt was correct by being able to see the most recently played cards.

Now need to start working more on the smooth movement of cards between piles etc