Fixed cards stacking in wrong direction since changing some stuff in the card piles. New variable in card piles for if should stack face up.
Card pile macro card rotation has been updated to choose between either stacking with the rotation face up or face down using the stack direction arrow now instead of the actual rotation of the card pile.

New bool variable in card piles for if we should stack the cards face up or not which is used in the card rotation macro.