Removed unused spheres and arrows from player. Card smooth movement.
New event in the card interface that replaces all instances of setting a card actor location and rotation.

Takes multiple parameters: new location (vector), new rotation (rotator), time to take (float), player camera (camera component), to attach to (scene component).

New variables in the card actor: location before moving, rotation before moving.

This event first sets these new variables to be the current location and rotation. Then sets the play rate of the upcoming timeline to be 1 / time to take.

Then we play the timeline from the start which just goes from 0 to 1 over 1 second smoothly which is used to lerp the rotator and location from the location and rotation before moving to the new location and rotation.
On finished we first check if the passed camera is valid and if so then we set the location and rotation again with the rotation being the camera's current world rotation and then we attach the card to the passed to attach to component. Then we call set relative location on hand add and then reset the previous do once.

This event replaces a lot of the previous instances of moving cards around.
Changed some stuff with removing from a pile so that it doesn't go to 0,0,0 and is hidden before moving and being unhidden so it is all smooth.

Had to change some stuff with the order of execution for the setup functions in the auction and idi game manager since moving the characters and then immediately setting up the card hands was glitching out so there's some delay between these now.

Also with the card hovering and selection whilst sorting had to change some stuff to have a delay before doing the hover check so that it is actually shown and completed properly.

Also for the update visible hand function had to make a new function for the hovering first card if no hovered cards found so that it can be called with a slight delay since we are moving the cards and this would reset the hovering.

Need to properly test all of this to make sure there are no bugs but yea now it kinda looks more like a game rather than a demo.