Rat screw rules setup. Rat screw moving both player and opponent properly now using target points.
Changed around and organised some stuff in the rat screw game manager.
Rules now shown properly after setting up rat screw.
New function in rat screw for starting the game which basically just calls the first turn function in the turn manager but is now called with a small delay when unpausing the game in the game mode. As per all of the other games.

Now using target points instead of blank actors for moving the characters in rat screw.
New functionality for moving the NPC in rat screw game manager which just does the same as the player but setting the actor rotation instead of the control rotation.

Setting control rotation as well as position for player in rat screw now as well.