Rules aesthetic update. Rules mapping context. Filled out i doubt it rules.
Changed the rules text holder to use a simple text box instead of a multi line one since it was kinda gray instead of white text.

Changed sizing for the rules text holders so more space for text.

New animation for the different rules widgets for increasing the blur over time so not instant. Functionality for calling playing this animation via an event in the corresponding widget which is called in the rules manager actor.

Rules manager actor now also gets the game mode and stores a reference to it. It uses this reference in the hide and show rules functions to add and remove the corresponding active game's mapping context and add or remove the new rules mapping context depending on if we are showing and hiding the rules.

The new rules mapping context only has the rules toggle input action on r so that we can't interact with any of the cards in the hand or do anything else whilst the rules are visible.

Filled out the I doubt it rules and the unique cards for I doubt it will do the other games tomorrow and will also setup a rules widget and functionality for darts as well tomorrow. Want to also try to merge the environment tomorrow possibly.