Small delay between showing rules and allowing rules to be hidden. Smoother card movement with sorting. Auction rules. Auction death unique change. Mini game rules. Darts rules button in mode selection widget. Rules delaying ability to close rules. Rules spamming fix. Stop widget interaction once started game with NPC.
There is now a variable delay between being able to see the rules and being able to close the rules so they have to be open for a small time which reduces the chance for weird bugs.
Also removed the self removal of the rules widget from parent in the widget event graph since it does basically get removed with the size box being set to scale 0,0.
This helped remove a bug with spamming the rules toggle button.

Smoother card movement with hovering and selecting when sorting now with some new events in the card interface for resetting the hovering and selecting timelines with these events being called when sorting has been completed and before the hover and select event calls.

Changed auction death unique card to deal all cards needs testing though but it should work.

Auction rules filled out.

Mini game rules setup.
New function in rules manager.
New button in darts mode selection for viewing rules.

Need to actually fill in the darts rules but that is easy.

When clicking start game now on a NPC we now stop the player movement with locking their move input which is then reset on the camera fade in (obviously it doesn't really matter since they can't move during the game but its just in case we forget for when the player returns after the game is over). Also on the click start game we also disable the ability for players to interact with other NPCs, this is done by setting can interact with NPC false for the player and also deactivating their widget interaction component.

New functionality in the player for closing the rules when not in game since this needs to be done for the mini games but we technically aren't in the game yet when viewing the rules with the rules button in the mini game mode widget.

Also the camera fades are smoother now for starting a game.