Angel doubting fix for IDI. Removed unused code for death in auction.
New function IDI game manager specifically for angel doubted.
This function moves the cards in the doubt pile back to the main pile and then calls the resume player turn function with half of the normal time.
Before the angel was just stopping the game from progressing if doubting it.

Tested death a little for auction think it works as expected now yayayay.