Filled out rat screw rules but no unique actions yet since not implemented yet.
Filled out the main rules of rat screw and the slapping stuff.
I'm not sure the slapping stuff is fully implemented since there was some slapping stuff with king then queen or vice versa that doesn't seem to be implemented yet so don't forget this as a task to do.
Not filled in the unique card stuff since they aren't implemented yet so don't forget to fill this in when finishing the unique card stuff for rat screw.