Removed unused code from player. Removed interact input action. Removed unused variables from player. Removed unused code from cardpile. File organisation.
Removed some unused code and variables from the player mainly the old NPC interaction system since it is now redundant and also had a reference to the NPC so we need to remove it for better memory management.
Also removed the associated input action for interacting with NPCs since it is now all with widgets anyway so can just use the click UI input action as has been standard for the last week or so.

Removed some old unused code from the card pile actor which was the old way of adding cards to the pile I think.

General file organisation to make some stuff cleaner and easier to navigate with some more subfolders.