Updated button prompt for darts. More animations for main menu and options menu. Pause menu changed some play animations to have finished events so can actually see the animations. Changed some of the NPC start game widget animations.
Button prompt for darts didnt have rules toggle now it does.

Shrink and enlarge animations for options instead of setting visibility on and off we now only ever set it visible and then call the shrink or enlarge animations in the options widget.

Click animations for the main menu setup and using play animation with finished event so can actually see the animation before the actual functionality is done.

Changed slightly some of the animations for the start game button widget for the NPCs.