Schrodingers Cat furthered possibly finished. Had to change some stuff with player collision but nothing major. Changed holo card collision so doesn't have any.
Schrodingers cat now has a timer when left line of sight which will be cleared if the cat enters line of sight again during the timer to give some leeway. Might need to change how to actually do the line of sight check since its a bit bad right now but its doable for now.
Using dot product of the (unit direction of the world location of the camera to the cat's location) and the forward vector of the camera then returning the inverse cos of the dot product and then checking if it is in a range from 0 to the FOV angle multiplied by 0.5 basically checking to see if it is viewed by the player's camera. We also check if the actor was recently rendered as well and if both of these checks are true then we do a line trace from the cat to the player to see if there is anything blocking and if there is then we know for certain that the cat is out of view and we will then move it and set it not active and deactivate the trace all with timers. If the cat is visible to the player then we set active true and stop the timers.

Changed the capsule collision on the player to overlap visibility trace so cat can line trace detect the player. Also changed the holo card to have no collision since it doesn't need any.