Testing cat. Audio and VFX Setup for cat. More functionality for stopping cat trace easier. Made 2 basic niagaras for the cat disappearing and reappearing.
Setup play sounds at location for cat moving.
Setup spawning emitters at location for cat moving.
More functionality for stopping the cat trace when needed by storing the reactivate cat trace after moving timer as a variable and also storing the time for reactivating the cat trace after moving as a float variable so more easily variable.

Think everything is working with the cat now.

Delay for cat being unhidden after moving so theres some disappearance and not just instantaneous moving.
Cat is now hidden when cat move is called.
Stop all timers function not currently used but will be in future when we want to disable the quantum properties of the cat.
Made 2 new niagara systems for the cat pop in and pop out of existence with the functionality to spawn these already implemented.
One is a simple burst from point and the other is a burst from point and then hanging particles.